voice-server/requirements.txt
setuptools==75.1.0
PyAudio==0.2.14
SpeechRecognition==3.10.4
requests
websocket-client==1.8.0
rel==0.4.9.19